minor improvements for the lists
